Rename tokenizer -> tokenize
